added / completed support to power measurement

+++ src/it/polito/elite/enocean/enj/eep/eep26/D2/D201/D201ActuatorMeasurementResponse.java
@@ -114,7 +114,7 @@ public class D201ActuatorMeasurementResponse
 	{
 		ByteBuffer buffer = ByteBuffer.wrap(measure);
 
-		return buffer.getDouble();
+		return (double)buffer.getInt();
 	}
 
 	/**
